@@ -2,13 +2,14 @@
   "name": "asl-puml",
   "version": "0.1.0",
   "description": "Generates a plant uml file from a valid ASL file",
-  "main": "./src/asl-puml.js",
+  "main": "./dist/asl-puml.js",
   "bin": {
-    "asl-puml": "./bin/asl-puml.js"
+    "asl-puml": "bin/asl-puml.ts"
   },
   "scripts": {
-    "lint": "./node_modules/eslint/bin/eslint.js . --max-warnings 0",
-    "test": "./node_modules/jest/bin/jest.js"
+    "build": "npx tsc",
+    "lint": "npx eslint . --max-warnings 0 --ext ts",
+    "test": "npm run build && npx jest"
   },
   "repository": {
     "type": "git",
@@ -30,18 +31,26 @@
   },
   "homepage": "https://github.com/massfords/asl-puml#readme",
   "devDependencies": {
-    "eslint": "^7.20.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.5",
-    "jest": "^26.6.3",
-    "jest-junit": "^12.0.0",
-    "prettier": "^2.7.1"
+    "@types/jest": "^28.1.3",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
+    "eslint": "^8.18.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-prettier": "^4.1.0",
+    "jest": "^28.1.1",
+    "jest-junit": "^14.0.0",
+    "prettier": "^2.7.1",
+    "ts-jest": "^28.0.5",
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
-    "asl-validator": "2.0.0",
+    "asl-validator": "^2.2.0",
     "commander": "^9.3.0",
-    "jsonpath": "^1.1.1",
     "jsonpath-plus": "^7.0.0"
   }
 }
